requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
bcrypt==4.0.1
pandas==2.1.4
openpyxl==3.1.2

# Servidores de produção (escolher um)
gunicorn==21.2.0  # Linux/Mac
waitress==2.1.2   # Windows
